added backward function and arrow hide function
Added the goBackward function to decrement the currCardIdx state by 1.
Updated the onClick event of the left arrow to call the goBackward function.
Added conditional rendering for the left and right arrows. The left arrow will only be displayed if the current card index is greater than 0, and the right arrow will only be displayed if the current card index is less than the total number of photos minus 1.